chore(app): moving back to amp but keeping the Unsafe api

+++ src/lib/Core/Directory.php
@@ -1,49 +1,57 @@
 <?php
 namespace CatPaw;
 
+use function Amp\async;
+use function Amp\File\createDirectory;
+use function Amp\File\createDirectoryRecursively;
+use function Amp\File\deleteDirectory;
+use function Amp\File\isDirectory;
+use function Amp\File\isFile;
+use function Amp\File\listFiles;
+use Amp\Future;
+
 use FilesystemIterator;
-use function React\Async\await;
-use React\Promise\Promise;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RecursiveRegexIterator;
 use RegexIterator;
+use SplFileInfo;
 use Throwable;
 
 class Directory {
     /**
-     * Delete a file or directory recursively.
-     * @param  string       $directoryName name of the directory.present.
+     * Delete a directory recursively.
+     * @param  string               $directoryName name of the directory.present.
      * @return Unsafe<void>
      */
     public static function delete(string $directoryName):Unsafe {
-        $directoryName = realpath($directoryName);
+        if (false === $directoryName) {
+            return error("Invalid directory $directoryName.");
+        }
+        
         try {
-            $directory = new RecursiveDirectoryIterator($directoryName);
-        } catch(Throwable $e) {
+            $list = listFiles($directoryName);
+        } catch (Throwable $e) {
             return error($e);
         }
 
-        $iterator = new RecursiveIteratorIterator($directory);
-
-        $directories = [];
-
-        for ($iterator->rewind();$iterator->valid();$iterator->next()) {
-            foreach ($iterator->current() as $fileName) {
-                if ($error = File::delete($fileName)->error) {
+        foreach ($list as $fileNameLocal) {
+            $fileName = "$directoryName/$fileNameLocal";
+            if (isFile($fileName)) {
+                if($error = File::delete($fileName)->error){
                     return error($error);
                 }
-                $currentDirectoryName = dirname($fileName);
-                if (!isset($directories[$currentDirectoryName])) {
-                    $directories[$currentDirectoryName] = true;
+            } else {
+                if($error = self::delete($fileName)->error){
+                    return error($error);
                 }
             }
         }
 
-        foreach (array_keys($directories) as $directoryName) {
-            if ($error = File::delete($directoryName)->error) {
-                return error($error);
-            }
+        try {
+            deleteDirectory($directoryName);
+        } catch (Throwable $e) {
+            return error($e);
         }
 
         return ok();
@@ -52,23 +60,12 @@ class Directory {
     /**
      * @return Unsafe<void>
      */
-    public static function create(string $directoryName):Unsafe {
-        if ('' === $directoryName) {
-            return error("Directory name cannot be empty.");
+    public static function create(string $directoryName, int $mode = 0777):Unsafe {
+        try {
+            return ok(createDirectoryRecursively($directoryName, $mode));
+        } catch (Throwable $e) {
+            return error($e);
         }
-
-        if (File::exists($directoryName)) {
-            if (!isDirectory($directoryName)) {
-                return error("The given directory $directoryName is actually an already existing file.");
-            }
-            return ok();
-        }
-
-        if (mkdir($directoryName, 0777, true)) {
-            return error("Could not create directory $directoryName.");
-        }
-
-        return ok();
     }
 
     /**
@@ -77,98 +74,69 @@ class Directory {
      * @param  string                $pattern       regex pattern to match while scanning.
      * @return Unsafe<array<string>>
      */
-    public static function flat(string $directoryName, false|string $pattern = false):Unsafe {
+    public static function flat(string $directoryName):Unsafe {
         try {
-            $directory = new RecursiveDirectoryIterator($directoryName);
+            $result = [];
+            $listAttempt = Directory::list($directoryName);
+            if($listAttempt->error){
+                return error($listAttempt->error);
+            }
+            foreach ($listAttempt->value as $fileName) {
+                if(isFile($fileName)){
+                    $result[] = $fileName;
+                } else {
+                    $flatAttempt = Directory::flat($fileName);
+                    if($flatAttempt->error){
+                        return error($flatAttempt->error);
+                    }
+                    $result = [...$result, ...$flatAttempt->value];
+                }
+            }
+            return ok($result);
         } catch(Throwable $e) {
             return error($e);
         }
-
-        $iterator = new RecursiveIteratorIterator($directory);
-
-        if (false !== $pattern) {
-            $iterator = new RegexIterator(
-                $iterator,
-                $pattern,
-                RecursiveRegexIterator::GET_MATCH
-            );
-        }
-
-        /** @var array<string> */
-        $fileNames = [];
-
-        for ($iterator->rewind();$iterator->valid();$iterator->next()) {
-            /** @var FileInfo */
-            $fileInfo = $iterator->current();
-            $fileName = $fileInfo->getFilename();
-            $filePath = $fileInfo->getRealPath();
-            if ('.' === $fileName || '..' === $fileName) {
-                continue;
-            }
-            $fileNames[] = realpath($filePath);
-        }
-
-        return ok($fileNames);
     }
 
     /**
      * List files and directories in a directory.
      * @param  string                $directoryName directory to scan.
-     * @param  string                $pattern       regex pattern to match while scanning.
      * @return Unsafe<array<string>>
      */
-    public static function list(string $directoryName, false|string $pattern = false):Unsafe {
+    public static function list(string $directoryName):Unsafe {
         try {
-            $iterator = new FilesystemIterator($directoryName);
-        } catch(Throwable $e) {
+            if(false === $directoryName){
+                return error("Directory $directoryName not found.");
+            }
+            $list = listFiles($directoryName);
+            $result = [];
+            foreach ($list as $fileName) {
+                $result[] = "$directoryName/$fileName";
+            }
+            return ok($result);
+        } catch (\Throwable $e) {
             return error($e);
         }
-
-        if (false !== $pattern) {
-            $iterator = new RegexIterator(
-                $iterator,
-                $pattern,
-                RecursiveRegexIterator::GET_MATCH
-            );
-        }
-
-        /** @var array<string> */
-        $fileNames = [];
-
-        for ($iterator->rewind();$iterator->valid();$iterator->next()) {
-            /** @var FileInfo */
-            $fileInfo = $iterator->current();
-            $fileName = $fileInfo->getFilename();
-            $filePath = $fileInfo->getRealPath();
-            if ('.' === $fileName || '..' === $fileName) {
-                continue;
-            }
-            $fileNames[] = realpath($filePath);
-        }
-
-        return ok($fileNames);
     }
 
 
     /**
      * Copy a directory.
-     * @param  string                $from
-     * @param  string                $to
-     * @param  string                $pattern regex pattern to match while scanning.
-     * @return Promise<Unsafe<void>>
+     * @param  string               $from
+     * @param  string               $to
+     * @param  string               $pattern regex pattern to match while scanning.
+     * @return Future<Unsafe<void>>
      */
-    function copy(string $from, string $to, false|string $pattern = false):Promise {
-        return new Promise(static function($ok) use ($from, $to, $pattern) {
+    function copy(string $from, string $to, false|string $pattern = false):Future {
+        return async(static function() use ($from, $to, $pattern) {
             if (!isDirectory($from)) {
-                $ok(error("Directory $from not found."));
-                return;
+                return error("Directory $from not found.");
             }
             
             try {
                 $iterator = new FilesystemIterator($from);
             } catch(Throwable $e) {
-                $ok(error($e));
-                return;
+                return error($e);
             }
 
             if (false !== $pattern) {
@@ -185,13 +153,12 @@ class Directory {
                 foreach ($iterator->current() as $fileName) {
                     $parts            = explode($key, $fileName, 2);
                     $relativeFileName = end($parts);
-                    if ($error = await(File::copy($fileName, "$to/$relativeFileName"))->error) {
-                        $ok(error($error));
-                        return;
+                    if ($error = File::copy($fileName, "$to/$relativeFileName")->await()->error) {
+                        return error($error);
                     }
                 }
             }
-            $ok(ok());
+            return ok();
         });
     }
